requirements.txt
# core
ujson
sqlalchemy
pandas
tables
scipy
richenum
matplotlib
fire
jinja2
sklearn
dulwich>=0.18.3

# optional
# pymysql

# For sphinx documentation only
sphinxcontrib-fulltoc
numpydoc
sphinx_rtd_theme

# For pypi document
pypandoc